+++ include/fc/crypto/sha224.hpp
@@ -19,7 +19,7 @@ class sha224
     char*    data()const;
     size_t data_size()const { return 224 / 8; }
 
-    static sha224 hash( const char* d, uint32_t dlen );
+    static sha224 hash( const char* d, size_t dlen );
     static sha224 hash( const string& );
 
     template<typename T>
@@ -36,7 +36,7 @@ class sha224
         encoder();
         ~encoder();
 
-        void write( const char* d, uint32_t dlen );
+        void write( const char* d, size_t dlen );
         void put( char c ) { write( &c, 1 ); }
         void reset();
         sha224 result();
