+++ src/crypto/hex.cpp
@@ -14,12 +14,12 @@ namespace fc {
       return 0;
     }
 
-    std::string to_hex( const char* d, uint32_t s ) 
+    std::string to_hex( const char* d, size_t s )
     {
         std::string r;
         const char* to_hex="0123456789abcdef";
         uint8_t* c = (uint8_t*)d;
-        for( uint32_t i = 0; i < s; ++i )
+        for( size_t i = 0; i < s; ++i )
             (r += to_hex[(c[i]>>4)]) += to_hex[(c[i] &0x0f)];
         return r;
     }
