+++ src/crypto/sha224.cpp
@@ -31,7 +31,7 @@ namespace fc {
       reset();
     }
 
-    sha224 sha224::hash( const char* d, uint32_t dlen ) {
+    sha224 sha224::hash( const char* d, size_t dlen ) {
       encoder e;
       e.write(d,dlen);
       return e.result();
@@ -40,7 +40,7 @@ namespace fc {
       return hash( s.c_str(), s.size() );
     }
 
-    void sha224::encoder::write( const char* d, uint32_t dlen ) {
+    void sha224::encoder::write( const char* d, size_t dlen ) {
       SHA224_Update( &my->ctx, d, dlen); 
     }
     sha224 sha224::encoder::result() {
